fix(core): break latent circular import between open_ai_types and datamodel
kiln_ai/utils/open_ai_types.py imported MessageUsage from
kiln_ai.datamodel.usage at module level. Reaching that submodule runs the
whole kiln_ai.datamodel package init, which chains through dataset_split ->
dataset_filters -> task_run, and task_run imports ChatCompletionMessageParam
and trace_has_pending_client_tool_calls back from open_ai_types -- names that
do not exist yet at that point. Importing kiln_ai.utils.open_ai_types as the
first kiln module raised ImportError.

The eager `from . import (...)` in adapters/__init__.py masked this by always
loading kiln_ai.datamodel first. Work in flight to make that init lazy (PEP
562 __getattr__, so multiprocessing.spawn children don't drag in litellm /
llama_index / numpy) turns the latent cycle into a hard failure, so the import
graph has to be correct first.

TYPE_CHECKING alone cannot fix this: open_ai_types builds a module-scope
TypeAdapter[list[ChatCompletionMessageParam]], which resolves the TypedDict
annotations at import time, so MessageUsage must be a real class.

Fix: move Usage/MessageUsage to kiln_ai/utils/usage.py, a leaf module that
imports nothing from kiln_ai at runtime, and re-export both from
kiln_ai/datamodel/usage.py. Every existing `from kiln_ai.datamodel.usage
import Usage, MessageUsage` site keeps working, as do pickles naming the old
module path. Kiln's own modules now import the canonical path so the shim
stays measurably external-only.

Verified unaffected by the move: no model_rebuild() or __module__-sensitive
code, no module-qualified Usage names in the generated OpenAPI schema or TS
client, and pdoc still builds.

Regression coverage in utils/test_import_layering.py. Rather than pinning one
case, it AST-scans datamodel for module-scope kiln_ai.utils imports (skipping
TYPE_CHECKING and function-local ones, which cannot cycle) and asserts in a
fresh subprocess that none of them pull in kiln_ai.datamodel -- so it keeps
covering new edges as the package grows. Its own scanner and probe-parsing are
unit tested, since a tripwire that quietly stops tripping is worse than none.

Usage class behaviour tests move to utils/test_usage.py beside the code;
datamodel/test_usage.py keeps the shim's back-compat contract.

Validated: importing either module first succeeds; with adapters/__init__.py
temporarily stubbed lazy (not committed), test_webhost.py and the full suite
pass; checks.sh exits 0.
