+++ libs/core/kiln_ai/utils/open_ai_types.py
@@ -32,7 +32,7 @@ from openai.types.chat.chat_completion_assistant_message_param import (
 )
 from typing_extensions import Required, TypedDict
 
-from kiln_ai.datamodel.usage import MessageUsage
+from kiln_ai.utils.usage import MessageUsage
 
 
 class ChatCompletionAssistantMessageParamWrapper(TypedDict, total=False):
