+++ libs/core/kiln_ai/utils/test_open_ai_types.py
@@ -12,7 +12,6 @@ from openai.types.chat import (
     ChatCompletionToolMessageParam as OpenAIChatCompletionToolMessageParam,
 )
 
-from kiln_ai.datamodel.usage import MessageUsage
 from kiln_ai.utils.open_ai_types import (
     KILN_ONLY_MESSAGE_FIELDS,
     ChatCompletionAssistantMessageParamWrapper,
@@ -23,6 +22,7 @@ from kiln_ai.utils.open_ai_types import (
 from kiln_ai.utils.open_ai_types import (
     ChatCompletionMessageParam as KilnChatCompletionMessageParam,
 )
+from kiln_ai.utils.usage import MessageUsage
 
 
 def test_assistant_message_param_properties_match():
